+++ libs/platform/src/lib/wizard-generator/wizard-generator.module.ts
@@ -5,6 +5,7 @@ import { DialogModule } from '@fundamental-ngx/core/dialog';
 import { ButtonModule } from '@fundamental-ngx/core/button';
 import { WizardModule } from '@fundamental-ngx/core/wizard';
 import { TitleModule } from '@fundamental-ngx/core/title';
+import { PlatformLinkModule } from '@fundamental-ngx/platform/link';
 import { LayoutGridModule } from '@fundamental-ngx/core/layout-grid';
 import { FormLabelModule } from '@fundamental-ngx/core/form';
 import { PlatformFormGeneratorModule } from '@fundamental-ngx/platform/form';
@@ -15,9 +16,10 @@ import { DialogWizardGeneratorComponent } from './components/dialog-wizard-gener
 import { WizardDialogGeneratorService } from './wizard-dialog-generator.service';
 import { WizardSummaryStepComponent } from './components/wizard-summary-step/wizard-summary-step.component';
 import { WizardSummarySectionComponent } from './components/wizard-summary-step/wizard-summary-section/wizard-summary-section.component';
-import { FdpWizardGeneratorGoNextButtonDirective } from './directives/fdp-wizard-generator-go-next-button.directive';
-import { FdpWizardGeneratorFinishButtonDirective } from './directives/fdp-wizard-generator-finish-button.directive';
+import { WizardGeneratorGoNextButtonDirective } from './directives/wizard-generator-go-next-button.directive';
+import { WizardGeneratorFinishButtonDirective } from './directives/wizard-generator-finish-button.directive';
 import { WizardGeneratorComponent } from './components/wizard-generator/wizard-generator.component';
+import { WizardGeneratorSummaryStepDirective } from './directives/wizard-generator-summary-step.directive';
 
 @NgModule({
     declarations: [
@@ -27,8 +29,9 @@ import { WizardGeneratorComponent } from './components/wizard-generator/wizard-g
         DialogWizardGeneratorComponent,
         WizardSummaryStepComponent,
         WizardSummarySectionComponent,
-        FdpWizardGeneratorGoNextButtonDirective,
-        FdpWizardGeneratorFinishButtonDirective
+        WizardGeneratorGoNextButtonDirective,
+        WizardGeneratorFinishButtonDirective,
+        WizardGeneratorSummaryStepDirective
     ],
     entryComponents: [DialogWizardGeneratorComponent],
     imports: [
@@ -40,7 +43,8 @@ import { WizardGeneratorComponent } from './components/wizard-generator/wizard-g
         PlatformButtonModule,
         LayoutGridModule,
         TitleModule,
-        FormLabelModule
+        FormLabelModule,
+        PlatformLinkModule
     ],
     exports: [
         WizardGeneratorComponent,
@@ -49,8 +53,9 @@ import { WizardGeneratorComponent } from './components/wizard-generator/wizard-g
         DialogWizardGeneratorComponent,
         WizardSummaryStepComponent,
         WizardSummarySectionComponent,
-        FdpWizardGeneratorGoNextButtonDirective,
-        FdpWizardGeneratorFinishButtonDirective
+        WizardGeneratorGoNextButtonDirective,
+        WizardGeneratorFinishButtonDirective,
+        WizardGeneratorSummaryStepDirective
     ],
     providers: [WizardDialogGeneratorService]
 })